Revert "engine"
This reverts commit 59c87123712bceff2249e96c8e564cf8700333bf.

+++ Engine/Node.cs
@@ -8,7 +8,7 @@ namespace Engine
 
         readonly int lng;
 
-        public Node(double[] coord)
+        public Node(decimal[] coord)
         {
             //transforms the latitude and longitude into integer numbers
             var correction = Convert.ToInt32(Math.Pow(10, 7));
